requirements.txt
Django==4.2
psycopg2-binary
dj-database-url
gunicorn
shortuuid
Pillow
djangorestframework
drf-nested-routers
markdown
django-filter